Moved cleanComponentTree funciton from its three file locations up the tree to the VisualContainer.tsx, the cleanedComponentTree is now passed down through props
