+++ src/app/Containers/VisualContainer.tsx
@@ -6,7 +6,7 @@ import Tree from '../components/StateTree/Tree';
 import Network from '../components/AtomNetwork/AtomNetwork';
 import AtomComponentVisualContainer from '../components/ComponentGraph/AtomComponentContainer';
 import Settings from '../components/Settings/SettingsContainer';
-import {stateSnapshot, selectedTypes} from '../../types';
+import {stateSnapshot, selectedTypes, componentAtomTree} from '../../types';
 // import Metrics from "../components/StateGraph/metrics";
 
 interface VisualContainerProps {
@@ -28,6 +28,52 @@ interface ZoomState {
 
 type navTypes = {
   [tabName: string]: JSX.Element;
+};
+
+const cleanComponentAtomTree = (
+  inputObj: componentAtomTree,
+): componentAtomTree => {
+  const obj = {} as componentAtomTree;
+  let counter = 0;
+  const innerClean = (inputObj: any, outputObj: any, counter: number = 0) => {
+    if (
+      inputObj.tag === 0 &&
+      inputObj.name !== 'RecoilRoot' &&
+      inputObj.name !== 'Batcher' &&
+      inputObj.name !== 'RecoilizeDebugger' &&
+      inputObj.name !== 'CssBaseline'
+    ) {
+      // if the obj is empty, we do this
+      if (Object.keys(obj).length === 0) {
+        outputObj.children = [];
+        outputObj.name = inputObj.name;
+        outputObj.recoilNodes = inputObj.recoilNodes;
+        outputObj.tag = inputObj.tag;
+        outputObj = outputObj.children;
+      }
+      // create another conditional
+      else {
+        const deepCopy: componentAtomTree = JSON.parse(
+          JSON.stringify(inputObj),
+        );
+        deepCopy.children = [];
+        outputObj.push(deepCopy);
+        if (outputObj.length > 1) {
+          outputObj = outputObj[outputObj.length - 1].children;
+        } else {
+          outputObj = outputObj[0].children;
+        }
+      }
+    }
+    // recursive call running through the whole component atom tree -- understand this better
+    for (let i = 0; i < inputObj.children.length; i++) {
+      innerClean(inputObj.children[i], outputObj, counter);
+    }
+    return outputObj;
+  };
+  innerClean(inputObj, obj, counter);
+  // returning the new object that we create
+  return obj;
 };
 
 // Renders Navbar and conditionally renders Diff, Visualizer, and Tree
@@ -60,6 +106,10 @@ const VisualContainer: React.FC<VisualContainerProps> = ({
   const componentAtomTree = currentSnapshot
     ? currentSnapshot.componentAtomTree
     : undefined;
+  const cleanedComponentAtomTree = currentSnapshot
+    ? cleanComponentAtomTree(currentSnapshot.componentAtomTree)
+    : undefined;
+
 
   // object containing all conditional renders based on navBar
   const nav: navTypes = {
@@ -76,6 +126,7 @@ const VisualContainer: React.FC<VisualContainerProps> = ({
     'Component Graph': (
       <AtomComponentVisualContainer
         componentAtomTree={componentAtomTree}
+        cleanedComponentAtomTree={cleanedComponentAtomTree}
         filteredCurSnap={filteredCurSnap}
         x={x}
         y={y}
@@ -88,7 +139,7 @@ const VisualContainer: React.FC<VisualContainerProps> = ({
     'Atom Network': <Network filteredCurSnap={filteredCurSnap} />,
 
     // individual snapshot visualizer
-    'Metrics': <Metrics componentAtomTree={componentAtomTree} />,
+    'Metrics': <Metrics cleanedComponentAtomTree={cleanedComponentAtomTree} />,
 
     // settings tab that doesn't want to be in quotes because too cool for school
     Settings: (
