+++ src/app/components/StateGraph/Visualizer.tsx
@@ -3,58 +3,12 @@ import * as d3 from 'd3';
 import {componentAtomTree} from '../../../types';
 
 interface VisualizerProps {
-  componentAtomTree: componentAtomTree;
+  cleanedComponentAtomTree: componentAtomTree;
 }
-// Recursive function that will run through componentatomtree, filter out unecessary nodes, and create the new object appropriately
-const cleanComponentAtomTree = (
-  inputObj: componentAtomTree,
-  ): componentAtomTree => {
-    const obj = {} as componentAtomTree;
-    let counter = 0;
-    const innerClean = (inputObj: any, outputObj: any, counter: number = 0) => {
-      if (
-        inputObj.tag === 0 &&
-        inputObj.name !== 'RecoilRoot' &&
-        inputObj.name !== 'Batcher' &&
-        inputObj.name !== 'RecoilizeDebugger' &&
-        inputObj.name !== 'CssBaseline'
-      ) {
-        // if the obj is empty, we do this
-        if (Object.keys(obj).length === 0) {
-          outputObj.children = [];
-          outputObj.name = inputObj.name;
-          outputObj.recoilNodes = inputObj.recoilNodes;
-          outputObj.tag = inputObj.tag;
-          outputObj = outputObj.children;
-        }
-        // create another conditional
-        else {
-          const deepCopy: componentAtomTree = JSON.parse(
-            JSON.stringify(inputObj),
-          );
-          deepCopy.children = [];
-          outputObj.push(deepCopy);
-          if (outputObj.length > 1) {
-            outputObj = outputObj[outputObj.length - 1].children;
-          } else {
-            outputObj = outputObj[0].children;
-          }
-        }
-      }
-      // recursive call running through the whole component atom tree -- understand this better
-      for (let i = 0; i < inputObj.children.length; i++) {
-        innerClean(inputObj.children[i], outputObj, counter);
-      }
-      return outputObj;
-    };
-    innerClean(inputObj, obj, counter);
-    // returning the new object that we create
-    return obj;
-  };
 
-const Visualizer: React.FC<VisualizerProps> = ({componentAtomTree}: any) => {
-  // data for bar graph render
-  const rawComponentAtomTree = cleanComponentAtomTree(componentAtomTree);
+
+const Visualizer: React.FC<VisualizerProps> = ({cleanedComponentAtomTree}: any) => {
+
   // create an empty array to store objects for property name and actualDuration
   let dataArray: {}[] = [];
   // function to traverse through the fiber tree
@@ -70,7 +24,7 @@ const Visualizer: React.FC<VisualizerProps> = ({componentAtomTree}: any) => {
     }
     node.children.forEach((child: any) => namesAndDurations(child))
   }
-  namesAndDurations(rawComponentAtomTree);
+  namesAndDurations(cleanedComponentAtomTree);
 
   const data = dataArray;
